refactor: move retrieval.py tests to test_retrieval.py
Move tests for BiEncoderModel.encode() and build_encoder_backbone()
to tests/unit_tests/_transformers/test_retrieval.py to match the
source module location. Keep LlamaBidirectionalModel-specific tests
in test_llama_bidirectional_model.py.
